tools/kubeconfig-merger/pyproject.toml
[project]
name = "kubeconfig-merger"
version = "1.0.0"
description = "Merge multiple kubeconfig files into a single output"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "kkamji", email = "rlaxowl5460@gmail.com" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Utilities",
]
requires-python = ">=3.9"
dependencies = ["PyYAML"]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "ruff>=0.8.0"]

[project.scripts]
kubeconfig-merger = "kubeconfig_merger.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
